Create config and utils files in server app, add missing MUI package
