+++ client/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.21",
     "@mui/material": "^5.15.20",
     "@mui/x-data-grid": "^7.7.0",
     "@reduxjs/toolkit": "^2.2.5",
