+++ server/src/config.ts
@@ -1,0 +1,11 @@
+import pg, { Pool } from "pg";
+
+const pool: Pool = new pg.Pool({
+  host: process.env.POSTGRES_HOST,
+  port: parseInt(process.env.POSTGRES_PORT || "5432"),
+  user: process.env.POSTGRES_USER,
+  password: process.env.POSTGRES_PASSWORD,
+  database: process.env.POSTGRES_DB,
+});
+
+export { pool };
